General

Description

Types

Simple

type Identifier = String

Identifies a value in the symbol table.

type ErrorMessage = String

Error messages thrown.

Execution Types

type Number = Double

The number type.

type Function

Arguments

= [Number]
The arguments given to the function.

-> Execution Number
The resulting action returning the function result.

The function type.

data Type

Union type of a NULL value, Numbers and Functions.

Constructors

Null
NULL value.

Number Number
A floating point number.

Function Function
A function mapping from a list of numbers to one number.

type Symbol = (Identifier, [Type])

An element of the symbol table, which is identified by a unique identifier and contains a stack of values of type Type, of which every element represents a variable identified by the identifier in a scope of the program.

Execution Monad

data ExitState

The final result of a script; it is either a Number or an Error message.

Constructors

Success Number
Successful computation returning a number.

Failure ErrorMessage
An error occured whilst execution.

type Execution a = EitherT ExitState (StateT [Symbol] IO) a

The Execution monad has a symbol table as its state and returns either an error message or a number.

run

Arguments

:: Execution ()
The program to run.

-> [Symbol]
The initial symbol table

-> IO ExitState
The result of an error message.

Executes the interpreted program.