Grammata.Execution

Description

Functions

Variables

declare

Arguments

:: Identifier
The identifier to be declared.

-> Execution ()
Resulting action.

If a symbol identified by the given identifier already exists, a new scope will be added, otherwise the a symbol will be added; however, it will be initiated with NULL.

assign

Arguments

:: Identifier
The identifier to which the number is to be assigned.

-> Expression
The Number to be assigned to the variable.

-> Execution ()
Resulting action.

Assignes the value of the given Number to the top legal scope of the symbol identified by the given identifier.

(.=)

Arguments

:: Identifier
The identifier to which the number is to be assigned.

-> Expression
The Number to be assigned to the variable.

-> Execution ()
Resulting action.

Infix version of assign

eval

Arguments

:: Expression
Expression to be evaluated.

-> Execution Number
Result of the evaluation.

Evaluation of arithmetical expressions.

Construction of functions

buildFunction

Arguments

:: Identifier
Identifier of the Function

-> [Identifier]
List of the function parameter names.

-> Execution ()
The body of the function.

-> Execution ()
The resulting action.

Builds the frame for a new function.

Program control

Control Structures

ifThenElse

Arguments

:: Expression
Condition

-> Execution ()
Action to be executed if the Condition is True.

-> Execution ()
Action to be executed if the Condition is False.

-> Execution ()
Resulting action.

An if .. then .. else .. statement

while

Arguments

:: Expression
Condition of continuation.

-> Execution ()
Execution to be iterated.

-> Execution ()
Resulting action.

A while loop

doWhile

Arguments

:: Expression
Condition of continuation.

-> Execution ()
Execution to be iterated.

-> Execution ()
Resulting action.

A do.. while loop

for

Arguments

:: Identifier
Counter variable

-> Expression
Stop value

-> Expression
Step size

-> Execution ()
Counter dependent action to be iterated

-> Execution ()
Resulting action.

A for loop

Termination

exitFailing

Arguments

:: ErrorMessage
The message returned on failure.

-> Execution a
The terminated action.

Terminates the execution with an error message.

exitSuccess

Arguments

:: Expression
The value returned on success.

-> Execution a
The terminated action.

Terminates the execution successfully returning a number.